requirements.txt
asgiref==3.4.1
Django==3.2
django-crispy-forms==1.13.0
django-rest-auth==0.9.5
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
django-allauth==0.48.0
natsort==8.1.0
